backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
langgraph
langchain
langchain-openai
langsmith